Crate bp_moonbase

Moonbase bridge primitives

Modules

betanet

Betanet bridge primitives

stagenet

Stagenet bridge primitives

Structs

BlockLength

Size limit of the Cumulus-based bridge hub blocks.

BlockWeights

Weight limit of the Cumulus-based bridge hub blocks.

Constants

AVERAGE_BLOCK_INTERVAL

Average block interval in Cumulus-based parachains.

BEST_FINALIZED_MOONBASE_WESTEND_HEADER_METHOD

Name of the <ThisChain>FinalityApi::best_finalized runtime method.

FREE_HEADERS_INTERVAL_FOR_MOONBASE_WESTEND_METHOD

Name of the <ThisChain>FinalityApi::free_headers_interval runtime method.

FROM_MOONBASE_WESTEND_MESSAGE_DETAILS_METHOD

Name of the From<ThisChain>InboundLaneApi::message_details runtime method.

MAX_UNCONFIRMED_MESSAGES_IN_CONFIRMATION_TX

Maximal number of unconfirmed messages at inbound lane for Cumulus-based parachains. Note: this value is security-relevant, decreasing it should not be done without careful analysis (like the one above).

MAX_UNREWARDED_RELAYERS_IN_CONFIRMATION_TX

Maximal number of unrewarded relayer entries at inbound lane for Cumulus-based parachains. Note: this value is security-relevant, decreasing it should not be done without careful analysis (like the one above).

TO_MOONBASE_WESTEND_MESSAGE_DETAILS_METHOD

Name of the To<ThisChain>OutboundLaneApi::message_details runtime method.

Traits

FromMoonbaseWestendInboundLaneApi

Inbound message lane API for messages sent by this chain.

MoonbaseWestendFinalityApi

API for querying information about the finalized chain headers.

ToMoonbaseWestendOutboundLaneApi

Outbound message lane API for messages that are sent to this chain.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance

Balance of an account.

BlockNumber

An index to a block.

Hash

A hash of some data used by the chain.

Hasher

The type of object that can produce hashes on Polkadot-like chains.

Header

Block header type as expected by this runtime.

LaneId

Bridge lane identifier.

Nonce

Nonce of a transaction on the Polkadot-like chains.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

Polkadot-like block signed with a Justification.