bp_moonbase
Module betanet

Betanet bridge primitives

Structs

Betanet

Betanet parachain.

Constants

PARACHAIN_ID

WITH_BRIDGE_MESSAGES_PALLET_NAME

Name of the messages pallet instance that is deployed at bridged chains.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance

Balance of an account.

BlockNumber

An index to a block.

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.